Rework for fsync-ba kernel
Move kernel-devel changes into concurrent pull/remove

Bump sigstore/cosign-installer from 3.5.0 to 3.6.0

Bumps [sigstore/cosign-installer](https://github.com/sigstore/cosign-installer) from 3.5.0 to 3.6.0.
- [Release notes](https://github.com/sigstore/cosign-installer/releases)
- [Commits](https://github.com/sigstore/cosign-installer/compare/v3.5.0...v3.6.0)

---
updated-dependencies:
- dependency-name: sigstore/cosign-installer
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

+++ build.sh
@@ -23,12 +23,15 @@ RELEASE="$(rpm -E %fedora)"
 # systemctl enable podman.socket
 
 # Install tmff2 driver.
-ln -s /usr/bin/ld.bfd /usr/bin/ld
+if [[ ! -f /usr/bin/ld ]]; then
+  ln -s /usr/bin/ld.bfd /usr/bin/ld
+fi
 mkdir -p /tmp/hid-tmff2
 cd /tmp/hid-tmff2
 git clone --recurse-submodules https://github.com/hellfur/hid-tmff2.git
 cd hid-tmff2
-CURRENT_BAZZITE_KERNEL=`ls /lib/modules`
+CURRENT_BAZZITE_KERNEL=$1 || `ls /lib/modules`
 make all KERNEL_VERSION=$CURRENT_BAZZITE_KERNEL
-sudo make install KERNEL_VERSION=$CURRENT_BAZZITE_KERNEL
+make install KERNEL_VERSION=$CURRENT_BAZZITE_KERNEL
 
+# rpm-ostree uninstall kernel-headers kernel-devel-$CURRENT_BAZZITE_KERNEL
